Move files that mostly help API requests to own directory
