+++ backend/api/routes/nonprofit.js
@@ -8,11 +8,11 @@ import {
   NonProfitAlreadyExistsError,
 } from "#errors/nonprofit-errors.js";
 import { createErrorReturn } from "#utils/validate-utils.js";
-import { generateStats } from "#utils/nonprofit-stat-utils.js";
+import { generateStats } from "#utils/api-helpers/nonprofit-stat-utils.js";
 import {
   checkNonProfitName,
   checkNonProfitId,
-} from "#utils/nonprofit-utils.js";
+} from "#utils/api-helpers/nonprofit-utils.js";
 import formatAddress from "#search/address-utils.js";
 import { checkLogin } from "#utils/session-utils.js";
