+++ backend/utils/api-helpers/service-utils.js
@@ -4,7 +4,7 @@ import { prisma } from "#utils/constants.js";
 // Local Imports
 import formatAddress from "#utils/search/address-utils.js";
 import { getAndValidateDate } from "#search/search-services.js";
-import { errorReturn, successReturn } from "./validate-utils.js";
+import { errorReturn, successReturn } from "../validate-utils.js";
 
 /**
  * Checks if a service exists with that name
